[feat] home : domain 구성(UseCase) 및 MVI 구현

+++ core/data/src/main/kotlin/com/twolskone/bakeroad/core/data/mapper/BakeryMapper.kt
@@ -8,7 +8,7 @@ internal fun RecommendBakeryResponse.toExternalModel(): RecommendBakery =
         id = bakeryId,
         name = name,
         rating = avgRating,
-        reviewCount = reviewCount,
+        reviewCount = reviewCount.toInt(),
         isOpened = isOpened,
         imageUrl = imgUrl
     )
